aruco-detector/requirements.txt
opencv-python
pillow
requests
numpy
